requirements.txt
ccxt>=4.3.0
pandas
pandas_ta
python-dotenv